misc: pricing units adjustments (#2338)
## Context

We are allowing customers to set custom pricing on the organization.
That means they can have this pricing unit displayed on invoices, and we
want for them to have a different display. This has already been
implemented for the vast majority of the cases

## Description

This PR is hence to adjust a little bit some UI elements to make sure we
match with the design and the expectations.

Fixes LAGO-900