misc: show PU Fiat amount value in drawer (#2341)
## Context

We have the ability to set a Pricing units on a charge hence having it
visible on the current usage.

## Description

We need to display the Fiat amount if the PU is displayed in the drawer.
Did that for the usage table but forgot to add it to the breakdown
drawer.
Also making sure the table lines' height is big enough to display the 2
lines in such case