feat: add billable metric limitation on wallet (#2357)
## Context

We would love to add a limitation on the wallet where you can only apply
the amount on certain fees related to certain billable metrics.

## Description

This pull request refactors the wallet form to add a new section.
This new section contains the expiration dates that were already present
on the different places of the form.
This new section also adds the possibility to select billable metrics on
your wallet, so the amount will be only scoped to them.

<!-- Linear link -->
Fixes LAGO-918