feat: CRUD for feature entitlement (#2352)
## Context

We are adding a new feature where people can create what we call
features that they can later assign on the plan. This will allow them to
control who has access to those plan features with some logic on top of
Lago.

## Description

This pull request is the first step of this feature. It adds all the
CRUD interactions such as:
- Feature lists
- Feature details
- Create and edit features form
- Allow the deletion of them

This functionality and its entry in the navigation bar are for now
hidden behind the frontend feature flag.

<!-- Linear link -->
Fixes LAGO-916