bug: prevent Table to call actionColumn wrongly (#2362)
While investigating a bug on the membership list, I noticed the root
cause was the Table calling a method `actionColumn` that is supposed to
have a real item passed in props.

However at this point the data was not present and this call just had
the same render each time -> a disabled button.

Just removing this call prevent deep object assertions to fail (e.g.
`membership.user.id`)

<!-- Linear link -->
Fixes ISSUE-1020