bug: make sure expected not found error don't raise alert (#2367)
## Context

We're fetching customer informations on the invoice preview, for display
purpose.

It can happen the customer is deleted hence cannot be fetched, resulting
in a not_found error and triggering an alert toast + sentry error.

## Description

This customer fetch is expected to potentially fetch and should not
raise "visual" error. We can then silent it!