misc: add beta chip on feature pages (#2369)
## Context

I'm working on a new feature "Feature and entitlement" and the List and
Form are already on main, hidden a feature flag.

We decided to flag this feature as `Beta` for it's firs first weeks.

## Description

This PR adds the `Beta` chips on those pages.

Also fix default spacing between elements in the page header layout.