fix: fetch finalized or draft invoices only (#2375)
## Context

Attempt to fix these kind of errors:
https://lago.sentry.io/issues/6581756497/events/7798dea65999490c80b47d48db013558/?project=6458937&referrer=replay-errors

## Description

In invoice list, we're fetching all invoices depending on their status
including generating invoices which could be the reason of `Resource not
found` type of error as the invoice is not completely "closed".

This is an attempt to resolve the issue by only fetching finalized+draft
invoices by default, and still be able to overwrite them with the
filters object.