misc: migrate some formatDateToTZ usage to intlFormatDateTime (#2377)
## Context

We've marked `formatDateToTZ` as deprecated and it should be replaced by
`intlFormatDateTime` method

## Description

This PR perform some changes but many remains.

It focuses on the "simple" ones that I could found. Those don't really
need special props and works out of the box with the default options of
`intlFormatDateTime`.

Later PRs will make sure to migrate the rest, had no time to keep on
with this task for now