feat: Regenerate invoice (#2291)
## Context

Create the new page for regenerating invoices.

## Description

Besides the page, this PR includes changes to existing logic, to re-use
as much as possible from what was already there. Mostly it's about how
the changes for regenerate invoices should be kept in a local state
(compared to draft invoices, where any fee changes trigger an API
request). To facilitate this, added `onAdd` / `onDelete` methods that
modify the local state instead of calling the backend.