feat: define boundaries on one-off fees (#2390)
## Context

We are allowing customers to set specific boundaries on fees when they
create a one-off invoice.

## Description

This pull request does two things:
1. It allows this definition through the invoice form
2. It removes some unused attributes from the query because it was too
complex
I also ensured the type used in the transform function was using the
fragment to make sure we are not missing any attributes when performing
those transformations.

<!-- Linear link -->
Fixes LAGO-945