feat: Projected usage (#2412)
## Context

We've decided to split the current usage query into two (default usage /
projected usage). This means that, based on the fact if projected usage
is enabled or not (premium integration), we call one of the two queries.

## Description

Re-added the changes we reverted. Most of the work went into fitting the
types together. At one point, in the future, we will re-join the
queries.