feat: assign entitlement to plan and subscriptions (#2416)
## Context

Our users want to determine when they can grant or revoke product
feature access to customers based on the plan they use. Some of those
features are not paying features, and Lago only lets you define charges
as part of a plan. There is no way to grant or revoke specific feature
access.

## Description

This pull request introduces the new functionality named 'features'.
This is a new entity you can assign to a plan or subscription to create
this new access logic.

<!-- Linear link -->
Fixes LAGO-917