fix: generate payment link tab (#2421)
## Context

On Safari, the copy to clipboard should follow a direct user
interaction.
But in this case we're waiting for the BE to answer with the call. It's
called transient user activation


## Description

When the url is fetched, open a new tab with the URL.
That will allow to manually copy the URL and fix the issue with Safari
users

I've also add an error catching in case we're flagging an error with the
PSP to display relevant toast message to the customer


Fixes issue-1044

---------

Co-authored-by: Alexandre Monjol <alexandre.monjol@gmail.com>