misc: extract fee display name logic (#2432)
## Context

The one time fee `displayName` logic is inlined in the template and
bring 2 issues
- Hard to extend and read
- Creates a linter warning

As I'm about to have this part of the app updated within a new feature,
it's time to refactor it a bit.

## Description

This PR extracts this logic into a method and makes it more predictable
and extensible.
it also removes the linter warning