bug: rename portuguese-brasil locale code (#2431)
## Context

A new `pt_BR` locale have been added recently.
We can save this informations in the customer, the billing entity and
the organization
It sets the locale for translations displayed in the transactional
emails and the customer portal.

However this locale does not follow the BCP 47 language syntax that
expect the language and the extended language tag to be separated by a
hyphen `-` instead of an underscore `_`.

## Description

This PR renames the locale to be `pr-BR` and updated the code
accordingly.
Also took the opportunity to improve typing around `DocumentLocales`
usage

<!-- Linear link -->
Fixes https://github.com/getlago/lago-front/issues/2420 and ISSUE-1048
Backend related PR: https://github.com/getlago/lago-api/pull/4233