bug: define default dateTo on some logs queries (#2442)
## Context

We have a bug when within a log list you fetch some collection, then a
lot of items are added in the database, and you keep scrolling the page.
The bug is about the fact that those new database elements will be
fetched and displayed after the old elements even though they are
earlier in the history.

This is due to our pagination system that is not using cursor-based
pagination but page-based pagination.

## Description

This pull request aims to fix this behavior by setting a default to date
with the value either to now or to the selected date by the customer.

<!-- Linear link -->
Fixes ISSUE-1071