misc(subscription): better overridden info display in list and details (#2452)
## Context

We're inconsistent in the way we display the info about an overridden
subscription.

We have this information displayed in the "general" subscription list
but not in the customer subscription list.

And this is not obvious what we call parent plan vs what is an
overriding plan for customers.

## Description

This pull request does:
- add the `Default/Overridden` column in the customer sub list
- Changes the `Parent plan` translation to `Override from` to make it
more explicit.

<!-- Linear link -->
Fixes ISSUE-1133