misc(creditNotes): validate form has at least one fee checked (#2453)
## Context

When creating a credit note, you can subit the form after unchecking all
fees.

Recently, Backend made a bugfix to return an error un such case, but
it's still allowed to submit the form like this and receive an error.

This form can accept 3 "shapes" of fee.
Classic ones for addOn and credit fees. And more "complex" ones that has
a different level of "checked" attribute.

## Description

This PR creates a check to prevent submitting the form if no fees are
selected.

Had to make a different function is used at the submit button level. As
it's checking on different attributes of the form, it was creating a
complex code there and was hard to test that way.
Preferred to extract the logic in the existing util and test the method
there.

<!-- Linear link -->
Fixes ISSUE-1108