bug: always hide entitlements on subscription details overview (#2454)
## Context

A plan can hold entitlements.
You can override this at the subscription level, and they are visible on
a dedicated tab, outside of the `Overview` details of the subscription,
that mainly display the plan's details.

Historically, this overview section uses the `EditPlan` fragment, that
todays fetches the plan's entitlements.
And we use the same components to display the plan details in the plan
details page and the subscription's>plan details.

So we have the "original" entitlements of the plan displayed in the
`Overview` tab, and the overriden ones in the `Entitlements` tab in the
subscription details.

## Description

This pull request adds a new attribute to this plan details components,
and you can now manually say if you want the entitlements section to be
hidden or not (default to shown).

So the subscription's plan overview can call it with
`showEntitlementSection={false}`

<!-- Linear link -->
Fixes ISSUE-1139