deps: bump cypress version to latest (#2464)
We have a [low severity
CVE](https://github.com/getlago/lago-front/security/dependabot/113) on a
sub deps of cypress.

It's fixed in their `15.0.0` so let's upgrade to the latest.

This `v15` brings the cypress studio and would be great to test it soon!

Other packages needs to be updated to have the CVE totally fixed on our
repo but they are not there yet.