feat(ai-agent): support graphql subscriptions (#2448)
## Context

Related to ai-agent feature.

This PR should update the instantiation of graphql to handle
subscriptions.

## Description

I added actioncable because that's the protocol we're using on the
backend side.
I've also split the `init` file to extract links for readability. 

As subscriptions are not used in the app yet, this can be safely merged