misc: compose charge accordion elements for better reusability (#2469)
## Context

I'm planning to have a whole new Charge accordion for fixed charges.

Those will have some differences with the event-based charges and today,
the only way to implement those change is either by making the charge
accordion API worst or make inner elements to be more "composable".

## Description

This PR starts to "compose" some components of the charge accordion, so
we can create this fixed charges accordion easier later.
The "main" benefit is visible in
`src/components/plans/UsageChargeAccordion.tsx` and in some create form
pages.

I've created a commit per split, and I'll create other PR as here I
think I only split half if the component.

It's a ready a again in readability and size (1116 lines to 739).