bug: restore wallet scope edition (#2472)
## Context

You can add some scope to the wallet when creating or editing it. This
scope is about either the type of fees it applies to or the billable
metric that can benefit from the wallet amount.

## Description

We currently have a bug where if none of the scopes of the billable
metric are defined, we don't send the applied-to. This leads to the fact
that if you want to delete the scope, it's never saved and sent to the
backend.

This pull request makes sure the 'applies to' is always centered back
and even if the arrays are empty, so we can reset the scope during a
wallet addition.

<!-- Linear link -->
Fixes ISSUE-1056