fix: Add missing premium integration flow for generating credit notes when regenerating (#2479)
## Context

The premium check flow was missing from the regenerate flow.

## Description

Decided to create a re-usable premium integration component, since it's
manually implemented in already a couple of places. Even if there's
already a <FreemiumAlert /> component, it's embedded deep in
CustomerRequestOverduePayment and it's only used in one place. That
component doesn't support sending emails, and it doesn't do dynamic
translations, so as part of a refactoring described in ISSUE-1193, we
could: 

- Replace `<FreemiumAlert />` with `<PremiumFeature />`
- Use it in:
    - ChargeBillingRadioGroup.tsx
    - WalletAccordion.tsx
    - SubscriptionUsageLifetimeGraph.tsx
    - ProgressiveBillingSection.tsx
    - SubscriptionCurrentUsageTable.tsx
    - SubscriptionAlertsList.tsx
    - RolePickerField.tsx
    - FeatureEntitlementSection.tsx
    - RotateApiKeyDialog.tsx
    - BillingEntityDunningCampaigns.tsx
    - Dunnings.tsx
    - FeaturesList.tsx
    - ApiKeysForm.tsx


<!-- Linear link -->
Fixes ISSUE-1192