bug: make sure the dev panel is closed on on orga switch (#2502)
## Context

You can open the developer panel and then switch organization.

In such cases, the developer panel remains open, and the problem is its
attempt to perform the call and sometimes with some ideas that are not
available on the new organization.
That results in the user receiving 404 errors and seeing an error toast.
The user could also land into an organization where his permission
doesn't allow him to see the developer panel.

## Description

This pull request simply closes the developer panel whenever the user
tries to switch organizations.

<!-- Linear link -->
Fixes ISSUE-1084