misc(auth): use new token if provided by Backend (#2505)
## Context

Our customers are "savagely" being kicked off the application after
three hours of usage of their authentication token.
We do not have any renew token logic in the application. However, the
backend does send us a new token in our responses header if the token
only has one hour remaining of activities allowed.

## Description

I'm just creating here a new Apollo link that is about to catch this
token, and if it's ever present in the header, it will replace the
current authentication token.

<!-- Linear link -->
Fixes ISSUE-978