bug(Charges): make sure options are reset based on choices (#2499)
## Context

We spotted 2 issues in the charge options settings
User are able to select some options when configuring their plan's
charges that are not supposed to happen.

1. `invoiceable` and `regroupPaidFees`
You have 2 places to define the `invoiceable` and `payInAdvance`
options.
`invoiceable` can be `true` or `false`
One of the options regarding invoicing methods allow to define
`regroupPaidFees` to `"invoice"`

That's kinda hard to explain by words here but the idea is that if you
select `payInAdvance: false` (pay in arrears), you should always have
`invoiceable: true` and `regroupPaidFees: null`


2. `minAmountCents` for invoices paid in advance.
If you define an invoice as to be paid in advance, it could be either
invoices now or later, but the invoice refers to the event you send and
there is no notion of minimum spending at this stage.
I noticed it as BE send us an error if you first save pay in arrears
with minimum amount cents, then edit the plan again and change to pay in
advance. The `minAmountCents` was not reset hence receiving an error.

## Description

This pull request contains 3 commits
1. File re-ordering
2. Fixes point `1.` of the context above
3. Fixes point `2.` of the context above

For the point `1.` from context, I preferred to fix it in the serializer
rather than when user clicks on the radio option.
I felt it was more natural for the user to first be able to select
different options, and if they didn't intend to reset the minimum
amounts, to keep the value there. So, it only resets when you click on
submitting the form.

## Media

|Pay in arrears|Pay in advance|
|-|-|
|<img width="658" height="299" alt="image"
src="https://github.com/user-attachments/assets/bb1086f5-d207-43fb-bbae-c486ee0f566a"
/>|<img width="652" height="323" alt="image"
src="https://github.com/user-attachments/assets/7d19b8b6-f3fd-4f0c-bb00-a0b74d90cbdb"
/>|


<!-- Linear link -->
Fixes ISSUE-1206