misc: add new formatAddress helper (#2427)
## Context

Today the address format follows the same pattern. Also the
implementation is very repetitive and verbose.

## Description

Let's use a lib to help with the address format and spread it's usage.
It simplifies a lot the logic around.
Address format is now decided based on the country code.

I'm waiting for the BE PR to merge

<!-- Linear link -->
Fixes ISSUE-975