misc(Charges): make the charge attribute on plan dynamic (#2552)
## Context

We are planning to add a new type of charge within the plan. However,
our inner components related to the charge type are very tied to the
`charge` attribute.
This pull request aims to make that attribute pointer dynamic so we can
later reuse the same components for this new charge type of the plan.

## Description

We now have a `chargeCursor` props in the `ChargeWrapperSwitch`
component.
It's passed down to children components and given to update methods
around, mostly.

I also had to rework some error handling in the `<ChargePercentage />`.