misc: extract formikProps from charge accordion elements (#2562)
This PR refactors the
`src/components/plans/chargeAccordion/ChargeWrapperSwitch.tsx` so it
does not pass the whole `formikProps` object down it's components.

It was kinda annoying me for a task I'm working on and it makes the code
simpler.
It's also a sort of first step to improve the perf of this form later.
We need to decouple the form from the components elements themself