bug: ensure luxon zone is UTC when running test (#2571)
I had issues running locally a test as I am currently in UTC -3. 

This pull request simply makes sure that every time the test runs, it
runs with the time zone UTC 0, ensuring that the expected data are
exactly the one that are printed here, no matter where you run the test.