misc: reorganise ChargeWrapperSwitch (#2570)
## Context

This pull request still represents preliminary work for the addition of
the new fixed charge type in the plan form.

## Description

It includes two commits that are quite self-explanatory from their
titles.

The first one moves a fragment from one component to another. 
It was initially located in `DynamicCharge` component, but since it
represents the pricing group keys, it now sits where it logically
belongs.

The second one looks more complex but is pretty straightforward. 
It shifts the pricing group key components and the new fixed charge
units one level higher within the `ChargeWrapperSwitch`.
This change required updating the DOM structure, adjusting class names,
and reworking the spacing between elements.

This is also the first pull request introducing the new fixed charge
elements into the codebase.
They might seem a bit odd for now, but they’re only here to enable
integration, for example, by declaring fragments or types that are used
elsewhere.
None of those fixed charge elements are visible to the user now.