fix: remove useless text in multicombo label (#2589)
## Context

While on a multicombo box that accept adding content (freeSolo option),
you can either `press enter` or `click` to add value. An example for
this is the zipcode filter in Customer.

If you press enter, everything work as intended. But if you click on the
visible option, the content of the Chip becomes `Click or press enter to
add "test"` instead of just "test".

This PR intends to fix this behaviour

## Description

I at first started to look at how the value is defined and tried to
redefine the onClick action but since it's coming from the MUI
Autocomplete, it was hard to change the onClick without potentially
creating side effects.

It's for this reason that I chose to modify the label of the Chip and
"filter out" the label if it is equal to the "add text" (translation key
`text_65ef30711cfd3e0083135de8`)

The fix is still up for debate, if anybody has another idea

<!-- Linear link -->
Fixes
[ISSUE-1209](https://linear.app/getlago/issue/ISSUE-1209/multiple-combobox-free-solo-adds-wrong-value-when-clicking-on-the)