fix: allow scroll in navigation tab on small screens (#2592)
## Context

On the Customer Details page, we have a very long navigation tab. On
small screns (tablet and smaller), we cannot scoll inside of it to
select the last elements

## Description

This PR fixes this issue by limiting the size of the element
encapsulating the navigation. This in turn allows for the overflows to
properly handle the scrolling

<!-- Linear link -->
Fixes
[ISSUE-993](https://linear.app/getlago/issue/ISSUE-993/navigationtab-is-broken-on-small-screens)