bug: fix array properties validations (#2599)
## Context

While working on the fixed charge addition, I noticed that our existing
validations are buggy and that they are not really testing the
validation schema for the charges.

## Description

What happened is that the property checked was not the good one and was
checking the singular attribute name instead of the plural one, which is
used in the application.
It was also asserting successful test as falsy, which makes the test
completely useless.

These pull requests ensure that all the cases are tested and that the
existing tests are passing with the new rules. There were also some gaps
in the validations, and I tried to make them more explicit within the
charge schema validation.

I also refactored this validation file to centralize the helper
functions, making the code clearer and easier to read.