misc: minor changes to prepare fixed charges (#2606)
## Context

I’m still preparing the codebase to integrate the new fixed charge into
the plans. It’s a big change, so I’m gradually introducing small,
non-breaking improvements to set things up. The idea is to reorganize
parts of the app now to reduce the scope and risk of the upcoming large
pull request.

## Description

This pull request implements few small changes, and are better to review
commit-by-commit.

There will be more to come, I'm trying to split the code as much as
possible and incrementally add them to the codebase.