misc: fix some linter warnings (#2635)
Removed some linter warning we had in the app

I tried as much as possible to preserve the logic to be equally the same
as before the refactoring. Otherwise, I took some time and the liberty
to refactor a bit.

When to each places to make sure it's not breaking anything