feat: add fixed charges (#2654)
## Context

All inner component are ready to handle the new fixed charges type (see
https://github.com/getlago/lago-front/pull/2653)

## Description

This PR now create the top level component that will take responsability
to display fixed charges in the plan/subscription details/form.

Once this PR and the one it relies on will be merged, the fixed charges
will be available in the app (behind a feature flag0

It does not mean the feature would be completely done, so expect other
PRs related to this topic.