feat: add fixed charges in plan and subscription
## Context

We’re introducing a new type of fixed charge associated with the plan.
These charges exist at the same level as the current ones but represent
a fixed cost that isn’t tied to any events.

## Description

This PR is pretty big, but I tried to split the changes per "domain".

Those commits should be reviewable independently and some are pretty
small.