misc: add missing copy and prevent future gap (#2662)
I noticed we have a missing translation in two of the "contextual"
locales, used in some parts of the app such as the customer portal.

This is not the first time so let's
- Add them
- Make sure it does not happen again using a script running on each PR

Please note that we exclude 2 locale files from this check:
- `base.json` as the whole application is not fully translated (yet)
- `pt-BR.json` as there were an attempt to translate the app in
Portuguese but have been abandoned so far. I'll soon take a decision
about it and clean the file / add it to this new check script