bug: hide premium feature upgrade for end users (#2670)
## Context

We display forecast usage in the sub usage tab.
This component is either rendered in the app or in the customer portal
(for end users).
It does not make sense to display it if the feature is not enabled, as
if it's shown, it contain a "upgrade" message anc CTA.

## Description

This PR allow to pass an attribute to this component, so we can hide the
tab if feature is not enabled AND the component is rendered in the
customer portal.

<!-- Linear link -->
Fixes ISSUE-1242