bug: make sure cache is up to date after evict (#2669)
## Context

When deleting an integration, we always evict the data from the cache.
When doing so, it happens that you land on the integration list and
create another integration. In such a case, the cache is still
considering the recently deleted data.
I believe this is a bug in Apollo GraphQL. However, I love to keep the
cache eviction logic in order to make the application more responsive.

## Description

A - not so nice -workaround for this issue is to make sure the cache is
up-to-date with the latest data after something has been evicted.
Doing so, it prevents Apollo from retrying an active query with a
wrong/unexisting data ID, leading to not show an error toast. anymore.

I've added this `refetchQueries` in the integration delete modal that
was missing them (some had it already)

<!-- Linear link -->
Fixes ISSUE-1240