bug: make sure entitlement is available for all plans (#2673)
## Context

We [recently](https://github.com/getlago/lago-front/pull/2593) made sure
that the feature entitlement is available for all plans, no matter if
you are premium or not.
A last check was missing, being removed in plan form.

## Description

This pull request makes sure that we don't display an upgrade model if
you are not a premium user for feature entitlements in Plan form.