feat(formatInvoice): fixed charges fees display order (#2674)
## Context

We'll soon release the fixed fee feature on the application, and we also
need to make sure that those are correctly displayed in the invoice
preview.

## Description

This pull request only updates the logic that defines in which order the
fees are placed on the invoice overview.
That now makes sure that fixed charge fees are displayed right after the
subscription ones.
I also added tests for the existing logic that was not covered. 

<!-- Linear link -->
Fixes LAGO-982