feat: include fixed fee in edit fee drawer combobox data (#2680)
## Context

As we’re introducing new fixed fees for plans and subscriptions, we also
need to support adding these fees in the invoice preview while the
invoice is in draft mode.
This addition takes place in the edit fee drawer, which should now
display both fixed charge fees and usage-based charge fees in the same
section.

## Description

This pull request ensures that we correctly build the combo box data
whenever the user needs to select a charge in the drawer.

The change may look a bit complex, but it essentially does two things:
1. Refactors the previous filter logic into a reduce method.
2. Adds another reduce method to handle fixed charges.
In the end, both arrays are combined to build the final combo box data.

You’ll also notice that the combo box data is now "grouped", including a
`group` property used to visually separate the two charge types in the
combo box.

The related tests have also been updated. 


This pull request lays the groundwork for fetching and injecting fixed
charges into the drawer, as well as adding more logic to it later on.
I decided to split this into a separate PR since the changes were
starting to get quite hard to follow.