bug: format email correctly on customer save (#2687)
## Context

We noticed a Sentry where a customer tried to save information with a
different format of email that the buyer can expect.

## Description

The frontend application should be responsible for sending the correct
data there, and the email should be a simple string separated by a comma
without any spaces in it.